+++ ChangeLog.md
@@ -1,11 +1,12 @@
 # Change Log #
 
 ## 0.38 ##
-_23 March 2022_
+_11 June 2022_
 - Updated formula for 8-bit H.264 content - changing from QSV using `--decode all --nvenc-gpu-only --nvenc-recommended --nvenc-cq 27 --rc-maxrate 15000`
 - Updated formula for software-based `x264` too:  `--x264-cbr --target 2160p=12000 --target 1080p=6000 --target 720p=3600 --target 480p=1800`
 - Swap around some of the menus to promote these two and demote QSV down as an optional transcode method. 
 - New folder for NVEnv H.264 post-processing added
+- Brought back Dropbox; dropped OneDrive
 
 
 ## 0.37 ##